+++ ProjectLex.InventoryManagement.Desktop/ViewModels/SupplierViewModels/SupplierListViewModel.cs
@@ -55,11 +55,15 @@ namespace ProjectLex.InventoryManagement.Desktop.ViewModels
 
         private void RemoveSupplier(SupplierViewModel supplierViewModel)
         {
-            _unitOfWork.SupplierRepository.Delete(supplierViewModel.Supplier);
-            _unitOfWork.Save();
-            _suppliers.Remove(supplierViewModel);
-            SupplierListViewHelper.RefreshCollection();
-            MessageBox.Show("Successful");
+            var result = MessageBox.Show("Do you really want to remove this item?", "Warning", MessageBoxButton.YesNo);
+            if (result == MessageBoxResult.Yes)
+            {
+                _unitOfWork.SupplierRepository.Delete(supplierViewModel.Supplier);
+                _unitOfWork.Save();
+                _suppliers.Remove(supplierViewModel);
+                SupplierListViewHelper.RefreshCollection();
+                MessageBox.Show("Successful");
+            }
         }
